Finish re-record option

+++ src/main/java/application/ui/pane/ListenPane.java
@@ -20,6 +20,7 @@ import org.json.simple.parser.JSONParser;
 import org.json.simple.parser.ParseException;
 import utils.buttons.Buttons;
 import utils.buttons.ImageButton;
+import utils.files.RecordVoice;
 
 import java.io.File;
 import java.io.FileReader;
@@ -30,6 +31,7 @@ public class ListenPane extends BorderPane {
 
     Main main;
     Stage primaryStage;
+    RecordVoice recordVoice;
     ResourceBundle language;
 
     HBox hbox;
@@ -56,17 +58,21 @@ public class ListenPane extends BorderPane {
     MediaPlayer[] listMediaPlayerAudioFile;
     Boolean[] listStatusMediaPlayerRecordFiles;
     Boolean[] listStatusMediaPlayerAudioFile;
+    Button[] listRecordingButton;
 
     int nbCorrespondingFile = 0;
     int nbPages;
     int actualPage = 1;
     int rowButtons = 12;
 
+    boolean runningRecord = false;
+
     public ListenPane(Main main, Stage primaryStage, ResourceBundle language){
         super();
 
         this.main = main;
         this.primaryStage = primaryStage;
+        this.recordVoice = new RecordVoice(main, main.getSaveFolder().getFolderPath());
         this.language = language;
 
         gridPane = new GridPane();
@@ -86,6 +92,7 @@ public class ListenPane extends BorderPane {
     public void setupListenPane(){
         this.nbCorrespondingFile = 0;
         this.actualPage = 1;
+        this.runningRecord = false;
         this.jsonPath = main.getSaveFolder().getJsonPath();
         this.recordPath = main.getSaveFolder().getRecordPath();
         this.getAllJsonFile();
@@ -179,6 +186,7 @@ public class ListenPane extends BorderPane {
         this.listStatusMediaPlayerRecordFiles = new Boolean[this.nbCorrespondingFile];
         this.listStartTimeAudioAndRecordFiles = new String[this.nbCorrespondingFile];
         this.listEndTimeAudioAndRecordFiles = new String[this.nbCorrespondingFile];
+        this.listRecordingButton = new Button[this.nbCorrespondingFile];
 
         this.filterNameList();
         this.initBooleanTab();
@@ -389,19 +397,36 @@ public class ListenPane extends BorderPane {
         reRecordButton.setPrefHeight(50);
         reRecordButton.setPrefWidth(300);
         reRecordButton.setOnAction((e) -> {
-            this.listMediaPlayerRecordFiles[indexTab].stop();
-            this.listMediaPlayerRecordFiles[indexTab].dispose();
-            String[] nameFile = listNameFilesCorrespondingToAudioFile[indexTab].split("\\.");
-            File recordNameFile = new File(recordPath + "/" + nameFile[0] + ".wav");
-            recordNameFile.delete();
-            this.newRecord();
-            this.reloadPage();
+            if (this.runningRecord){
+                this.newRecord(false, indexTab);
+            }else {
+                this.runningRecord = true;
+                this.listMediaPlayerRecordFiles[indexTab].stop();
+                this.listMediaPlayerRecordFiles[indexTab].dispose();
+                String[] nameFile = listNameFilesCorrespondingToAudioFile[indexTab].split("\\.");
+                File recordNameFile = new File(recordPath + "/" + nameFile[0] + ".wav");
+                recordNameFile.delete();
+                ((ImageView) reRecordButton.getGraphic()).setImage(new Image("images/stopReRecord.png"));
+                this.newRecord(true, indexTab);
+            }
         });
-        this.gridPane.add(reRecordButton, 6,indexGridPane+1);
+        this.listRecordingButton[indexTab] = reRecordButton;
+        this.gridPane.add(reRecordButton,6,indexGridPane+1);
     }
 
-    public void newRecord(){
-
+    public void newRecord(boolean start, int indexTab){
+        if (start){
+            for (Button button : this.listRecordingButton){
+                button.setDisable(true);
+            }
+            this.listRecordingButton[indexTab].setDisable(false);
+            this.recordVoice.startRecording();
+        }else {
+            this.recordVoice.stopRecording();
+            String[] nameFile = listNameFilesCorrespondingToAudioFile[indexTab].split("\\.");
+            this.recordVoice.renameTempAudioFile(nameFile[0]);
+            this.reloadPage();
+        }
     }
 
     public void clearGridPane(){
